target/python/4_0_0/pyproject.toml
[build-system]
requires = ["setuptools>=62.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-cdm"
version = "0.0.0"
requires-python = ">= 3.10"
dependencies = [
   "pydantic",
   "rosetta.runtime==1.0.0"
]
[tool.setuptools.packages.find]
where = ["src"]